+++ src/routes/game/play/Play.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Play from './play';
+import Play from './Play';
 import { renderWithProviders } from 'utils/TestUtils';
 
 it('renders without crashing', () => {
